refactor: rewrite imageGenerator to ts
