+++ package.json
@@ -33,6 +33,8 @@
   "packageManager": "pnpm@9.15.9+sha512.68046141893c66fad01c079231128e9afb89ef87e2691d69e4d40eee228988295fd4682181bae55b58418c3a253bde65a505ec7c5f9403ece5cc3cd37dcf2531",
   "devDependencies": {
     "@rslib/core": "^0.6.3",
+    "@types/fs-extra": "^11.0.4",
+    "@types/node": "^22.14.1",
     "prettier": "^3.5.3"
   },
   "engines": {
