+++ src/index.js
@@ -4,7 +4,7 @@
  */
 import fs from 'fs-extra';
 import path from 'path';
-import { generateFullPageImages } from './imageGenerator.js';
+import { generateFullPageImages } from './imageGenerator';
 import { ensureDir, removeFile, extractMdFromLLMOutput, adjustMarkdownHeadings, getOldMarkdownHeadings } from './utils.js';
 import ModelClient from './modelClient.js';
